requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues
azure.storage.blob
azure-functions
requests
boto3
botocore
python-dateutil
setuptools
validators
jmespath
certifi
six
urllib3
idna
packaging
pyparsing
wheel
msgspec
configparser
azure-identity
azure-keyvault-secrets